assign import as a component
