+++ src/Components/index.tsx
@@ -2,3 +2,4 @@ export { default as Home } from './Home/Home';
 export { default as CreateVideo } from './CreateVideo/CreateVideo';
 export { Player } from './Import/Player';
 export { Modal } from './Modal/Modal';
+export {isImage, isVideo} from './utils';
